% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARIS_home.R
\name{BARIS_home}
\alias{BARIS_home}
\title{Home Page Datasets}
\usage{
BARIS_home()
}
\value{
A dataframe with datasets characteristics
}
\description{
This function displays the most important information about the datasets that are currently displayed within the home page of the data.gouv portal.
}
\examples{
\dontrun{

BARIS_home()}
}
