% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARIS_resources.R
\name{BARIS_resources}
\alias{BARIS_resources}
\title{Displaying resources within a data set}
\usage{
BARIS_resources(datasetId)
}
\arguments{
\item{datasetId}{the unique identifier of a data set}
}
\value{
a table of all the data frames (resources) included within a specific data set and their main characteristics
}
\description{
The data.gouv API provides access to several data sets in which one can find several individual data frames to exploit. The BARIS_resources lists all the data frames available within a specific data set along with the main information concerning the data frames.
}
\examples{
\donttest{

BARIS_resources("53699934a3a729239d2051a1")}
}
