% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARIS_search.R
\name{BARIS_search}
\alias{BARIS_search}
\title{Searching for a data set}
\usage{
BARIS_search(query, page_number = 1, page_size = 20)
}
\arguments{
\item{query}{a character string defining the research.}

\item{page_number}{numeric value denoting which page to search for, defaults to 1}

\item{page_size}{numeric value, for a specific page (see \code{page_number}), how
many results do you want to query. Defaults to 20}
}
\value{
a character string
}
\description{
This function searches specific data sets through the data.gouv API according to the pattern provided into the query argument.
}
\examples{
\donttest{

BARIS_search("Marseille")}
}
