% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARIS_explain.R
\name{BARIS_explain}
\alias{BARIS_explain}
\title{Data set description}
\usage{
BARIS_explain(datasetId)
}
\arguments{
\item{datasetId}{the unique id of a data set}
}
\value{
a character string describing a data set
}
\description{
This function provides in French a detailed description of a data set.
}
\examples{
\donttest{

BARIS_explain("53699934a3a729239d2051a1")}
}
