% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BARIS_search.R
\name{BARIS_search}
\alias{BARIS_search}
\title{Searching for a dataset}
\usage{
BARIS_search(query, n_pages = 20)
}
\arguments{
\item{query}{a character string through which to search}

\item{n_pages}{number of pages to consider when searching (default to 20). For more, information check out the pagination section of the data.gouv API documentation}
}
\value{
a character string
}
\description{
This function searches specific datasets through the data.gouv API according to the pattern provided into the query argument.
}
\examples{
\dontrun{

BARIS_search("Marseille")}
}
